+++ freemarker-generator-cli/src/main/java/org/apache/freemarker/generator/cli/task/FreeMarkerTask.java
@@ -18,15 +18,22 @@ package org.apache.freemarker.generator.cli.task;
 
 import freemarker.template.Configuration;
 import freemarker.template.Template;
+import freemarker.template.TemplateException;
 import org.apache.commons.io.FileUtils;
 import org.apache.freemarker.generator.base.FreeMarkerConstants.Location;
 import org.apache.freemarker.generator.base.datasource.DataSource;
 import org.apache.freemarker.generator.base.datasource.DataSourceFactory;
 import org.apache.freemarker.generator.base.datasource.DataSources;
+import org.apache.freemarker.generator.base.template.TemplateOutput;
+import org.apache.freemarker.generator.base.template.TemplateSource;
+import org.apache.freemarker.generator.base.template.TemplateTransformation;
+import org.apache.freemarker.generator.base.template.TemplateTransformations;
 import org.apache.freemarker.generator.base.util.UriUtils;
 import org.apache.freemarker.generator.cli.config.Settings;
 
+import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Writer;
 import java.net.URI;
@@ -47,6 +54,7 @@ import static org.apache.freemarker.generator.cli.config.Suppliers.configuration
 import static org.apache.freemarker.generator.cli.config.Suppliers.dataModelSupplier;
 import static org.apache.freemarker.generator.cli.config.Suppliers.dataSourcesSupplier;
 import static org.apache.freemarker.generator.cli.config.Suppliers.parameterSupplier;
+import static org.apache.freemarker.generator.cli.config.Suppliers.templateTransformationsSupplier;
 import static org.apache.freemarker.generator.cli.config.Suppliers.toolsSupplier;
 
 /**
@@ -62,41 +70,60 @@ public class FreeMarkerTask implements Callable<Integer> {
     private final Supplier<Map<String, Object>> dataModelsSupplier;
     private final Supplier<Map<String, Object>> parameterModelSupplier;
     private final Supplier<Configuration> configurationSupplier;
+    private final Supplier<TemplateTransformations> templateTransformationsSupplier;
+
 
     public FreeMarkerTask(Settings settings) {
         this(settings,
-                toolsSupplier(settings),
+                configurationSupplier(settings),
+                templateTransformationsSupplier(settings),
                 dataSourcesSupplier(settings),
                 dataModelSupplier(settings),
                 parameterSupplier(settings),
-                configurationSupplier(settings));
+                toolsSupplier(settings)
+        );
     }
 
     public FreeMarkerTask(Settings settings,
-                          Supplier<Map<String, Object>> toolsSupplier,
+                          Supplier<Configuration> configurationSupplier,
+                          Supplier<TemplateTransformations> templateTransformationsSupplier,
                           Supplier<List<DataSource>> dataSourcesSupplier,
                           Supplier<Map<String, Object>> dataModelsSupplier,
                           Supplier<Map<String, Object>> parameterModelSupplier,
-                          Supplier<Configuration> configurationSupplier) {
+                          Supplier<Map<String, Object>> toolsSupplier) {
         this.settings = requireNonNull(settings);
         this.toolsSupplier = requireNonNull(toolsSupplier);
         this.dataSourcesSupplier = requireNonNull(dataSourcesSupplier);
         this.dataModelsSupplier = requireNonNull(dataModelsSupplier);
         this.parameterModelSupplier = requireNonNull(parameterModelSupplier);
         this.configurationSupplier = requireNonNull(configurationSupplier);
+        this.templateTransformationsSupplier = requireNonNull(templateTransformationsSupplier);
     }
 
     @Override
     public Integer call() {
-        final Template template = template(settings, configurationSupplier);
-        try (Writer writer = settings.getWriter(); DataSources dataSources = dataSources(settings, dataSourcesSupplier)) {
+        try {
+            final Configuration configuration = configurationSupplier.get();
+            final TemplateTransformations templateTransformations = templateTransformationsSupplier.get();
+            final DataSources dataSources = dataSources(settings, dataSourcesSupplier);
             final Map<String, Object> dataModel = dataModel(dataSources, parameterModelSupplier, dataModelsSupplier, toolsSupplier);
-            template.process(dataModel, writer);
+            templateTransformations.getList().forEach(t -> process(configuration, t, dataModel));
             return SUCCESS;
         } catch (RuntimeException e) {
-            throw e;
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to render FreeMarker template: " + template.getName(), e);
+            throw new RuntimeException("Failed to process templates", e);
+        }
+    }
+
+    private void process(Configuration configuration,
+                         TemplateTransformation templateTransformation,
+                         Map<String, Object> dataModel) {
+        final TemplateSource templateSource = templateTransformation.getTemplateSource();
+        final TemplateOutput templateOutput = templateTransformation.getTemplateOutput();
+        try (Writer writer = writer(templateOutput)) {
+            final Template template = template(configuration, templateSource);
+            template.process(dataModel, writer);
+        } catch (TemplateException | IOException e) {
+            throw new RuntimeException("Failed to process template: " + templateSource.getName(), e);
         }
     }
 
@@ -113,33 +140,6 @@ public class FreeMarkerTask implements Callable<Integer> {
         return new DataSources(dataSources);
     }
 
-    /**
-     * Loading FreeMarker templates from absolute paths is not encouraged due to security
-     * concern (see https://freemarker.apache.org/docs/pgui_config_templateloading.html#autoid_42)
-     * which are mostly irrelevant when running on the command line. So we resolve the absolute file
-     * instead of relying on existing template loaders.
-     *
-     * @param settings              Settings
-     * @param configurationSupplier Supplies FreeMarker configuration
-     * @return FreeMarker template
-     */
-    private static Template template(Settings settings, Supplier<Configuration> configurationSupplier) {
-        final String templateName = settings.getTemplateName();
-        final Configuration configuration = configurationSupplier.get();
-
-        try {
-            if (settings.isInteractiveTemplate()) {
-                return interactiveTemplate(settings, configuration);
-            } else if (isAbsoluteTemplateFile(settings)) {
-                return fileTemplate(settings, configuration);
-            } else {
-                return configuration.getTemplate(templateName);
-            }
-        } catch (IOException e) {
-            throw new RuntimeException("Failed to load template: " + templateName, e);
-        }
-    }
-
     private static Map<String, Object> dataModel(
             DataSources dataSources,
             Supplier<Map<String, Object>> parameterModelSupplier,
@@ -153,27 +153,55 @@ public class FreeMarkerTask implements Callable<Integer> {
         return result;
     }
 
-    private static boolean isAbsoluteTemplateFile(Settings settings) {
-        final File file = new File(settings.getTemplateName());
-        return file.isAbsolute() && file.exists() & !file.isDirectory();
+    // ==============================================================
+
+    private static Writer writer(TemplateOutput templateOutput) throws IOException {
+        if (templateOutput.getWriter() != null) {
+            return templateOutput.getWriter();
+        }
+
+        final File file = templateOutput.getFile();
+        FileUtils.forceMkdirParent(file);
+        return new BufferedWriter(new FileWriter(file));
     }
 
-    private static Template fileTemplate(Settings settings, Configuration configuration) {
-        final String templateName = settings.getTemplateName();
-        final File templateFile = new File(templateName);
-        try {
-            final String content = FileUtils.readFileToString(templateFile, settings.getTemplateEncoding());
-            return new Template(templateName, content, configuration);
-        } catch (IOException e) {
-            throw new RuntimeException("Failed to load template: " + templateName, e);
+    /**
+     * Loading FreeMarker templates from absolute paths is not encouraged due to security
+     * concern (see https://freemarker.apache.org/docs/pgui_config_templateloading.html#autoid_42)
+     * which are mostly irrelevant when running on the command line. So we resolve the absolute file
+     * instead of relying on existing template loaders.
+     *
+     * @param configuration  FreeMarker configuration
+     * @param templateSource source template to load
+     * @return FreeMarker template
+     */
+    private static Template template(Configuration configuration, TemplateSource templateSource) {
+        switch (templateSource.getOrigin()) {
+            case PATH:
+                return fromTemplatePath(configuration, templateSource);
+            case CODE:
+                return fromTemplateCode(configuration, templateSource);
+            default:
+                throw new IllegalArgumentException("Don't know how to handle: " + templateSource.getOrigin());
         }
     }
 
-    private static Template interactiveTemplate(Settings settings, Configuration configuration) {
+    private static Template fromTemplatePath(Configuration configuration, TemplateSource templateSource) {
+        final String path = templateSource.getPath();
         try {
-            return new Template(Location.INTERACTIVE, settings.getInteractiveTemplate(), configuration);
+            return configuration.getTemplate(path);
         } catch (IOException e) {
-            throw new RuntimeException("Failed to load interactive template", e);
+            throw new RuntimeException("Failed to load template from path: " + path, e);
+        }
+    }
+
+    private static Template fromTemplateCode(Configuration configuration, TemplateSource templateSource) {
+        final String name = templateSource.getName();
+
+        try {
+            return new Template(name, templateSource.getCode(), configuration);
+        } catch (IOException e) {
+            throw new RuntimeException("Failed to load template code: " + name, e);
         }
     }
 }